Implement Feature Request #5 from Deerhound579/mcgill-course-map
This patch renamed the ratio button "Path by course" and added a new one "Prerequisites that you need for this course."

Note that this patch introduces a bug to the program. After switching to the `overview` mode using the ratio buttons, it is no longer possible to switch back to `course` or `prerequisite`. Please look into this issue.