Implement visualization of node depths
In graph.py, the function calculate_depth finds the maximum distance from each node to a leaf.

In app.py, the get_color_coded_style_sheet function adds a few color-coded selectors to overview_stylesheet (and return the updated stylesheet.)

Current issue: the stylesheet won't update when the user enters a new query. However, it should, such that the colors are assigned based on the max_depth of this particular query.